requirements.txt
attrs
beautifulsoup4
lxml
wget
pymongo
aiohttp
json5
pytest
pycld3
pycountry
click
jsonlines

spacy
pythainlp
amseg
parsivar
razdel
laonlp
khmer-nltk
stanza
spacy-pkuseg>=0.0.27,<0.1.0